% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.VARCP}
\alias{plot.VARCP}
\title{Plots estimated test statistic}
\usage{
\method{plot}{VARCP}(VARCP, a = 0.05, ...)
}
\arguments{
\item{VARCP}{A list of class "VARCP"}

\item{a}{Level of significance for which critical values are computed}

\item{...}{Additional graphical arguments}
}
\value{
Displays a plot with automaticaly computed critical values
}
\description{
This function creates plot of computed test statistic with ability to add critical values as a red line
}
\examples{
## Simulation of time series of length 200
phi <- matrix(c(0.2,-0.6,0.3,1),2,2)
sig <- matrix(c(4,0.8,0.8,1),2,2)
data <- VARMAsim(200, arlags = 1, phi = phi, sigma = sig)
model <- VAR(data[["series"]], p = 1)

## Test estimation
CUSUM <- statcomp(model, type = "CUSUM")

Visualizing computed test statistic
plot(CUSUM, a = 0.1)
plot(CUSUM, a = 0.05)

}
