% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcomp.R
\name{statcomp}
\alias{statcomp}
\title{Computes structural change test statistic}
\usage{
statcomp(model, type = "CUSUM", trim = NULL)
}
\arguments{
\item{model}{A list returned by VAR or VARMA functions from MTS package}

\item{type}{A character string which determines type of the test statistic returned by the function, either "CUSUM", "LRT" or "DET". The default is "CUSUM"}

\item{trim}{A number of time points on data edges for which test statistic is not computed}
}
\value{
statcomp returns a list of class "VARCP" with following items
\item{values}{Named vector of values of selected test statistic, where name is time index for which statistic is computed}
\item{type}{Character string which states type of the computed test statistic}
}
\description{
Computes test statistic value for each time point.
Likelihood-ratio test (LRT), Cumulative sum test (CUSUM) and Darling-Erdös-type test (DET) statistics are supported.
}
\examples{
## Simulation of time series of length 200
phi <- matrix(c(0.2,-0.6,0.3,1),2,2)
sig <- matrix(c(4,0.8,0.8,1),2,2)
data <- VARMAsim(200, arlags = 1, phi = phi, sigma = sig)
model <- VAR(data[["series"]], p = 1)

## Tests estimation
CUSUM <- statcomp(model, type = "CUSUM")
LRT <- statcomp(model, type = "LRT", trim = 15)

}
